@@ -2,8 +2,6 @@ package com.example.android.recyclerplayground.layout;
 
 import android.content.Context;
 import android.graphics.PointF;
-import android.support.v7.widget.LinearSmoothScroller;
-import android.support.v7.widget.RecyclerView;
 import android.util.AttributeSet;
 import android.util.Log;
 import android.util.SparseArray;
@@ -11,11 +9,15 @@ import android.util.SparseIntArray;
 import android.view.View;
 import android.view.ViewGroup;
 
+import androidx.recyclerview.widget.LinearSmoothScroller;
+import androidx.recyclerview.widget.RecyclerView;
+
 import java.util.HashSet;
 import java.util.List;
 
+
 /**
- * A {@link android.support.v7.widget.RecyclerView.LayoutManager} implementation
+ * A {@link androidx.recyclerview.widget.RecyclerView.LayoutManager} implementation
  * that places children in a two-dimensional grid, sized to a fixed column count
  * value. User scrolling is possible in both horizontal and vertical directions
  * to view the data set.
@@ -26,8 +28,8 @@ import java.util.List;
  *
  * <p>This manager does make some assumptions to simplify the implementation:
  * <ul>
- *     <li>All child views are assumed to be the same size</li>
- *     <li>The window of visible views is a constant</li>
+ * <li>All child views are assumed to be the same size</li>
+ * <li>The window of visible views is a constant</li>
  * </ul>
  */
 public class FixedGridLayoutManager extends RecyclerView.LayoutManager {
@@ -65,6 +67,7 @@ public class FixedGridLayoutManager extends RecyclerView.LayoutManager {
     /**
      * Set the number of columns the layout manager will use. This will
      * trigger a layout update.
+     *
      * @param count Number of columns.
      */
     public void setTotalColumnCount(int count) {
@@ -156,7 +159,7 @@ public class FixedGridLayoutManager extends RecyclerView.LayoutManager {
          */
         if (state.isPreLayout()) {
             removedCache = new SparseIntArray(getChildCount());
-            for (int i=0; i < getChildCount(); i++) {
+            for (int i = 0; i < getChildCount(); i++) {
                 final View view = getChildAt(i);
                 LayoutParams lp = (LayoutParams) view.getLayoutParams();
 
@@ -218,10 +221,10 @@ public class FixedGridLayoutManager extends RecyclerView.LayoutManager {
              * is much smaller than it was before, and you are scrolled to
              * a location where no items would exist.
              */
-            int maxFirstRow = getTotalRowCount() - (mVisibleRowCount-1);
-            int maxFirstCol = getTotalColumnCount() - (mVisibleColumnCount-1);
+            int maxFirstRow = getTotalRowCount() - (mVisibleRowCount - 1);
+            int maxFirstCol = getTotalColumnCount() - (mVisibleColumnCount - 1);
             boolean isOutOfRowBounds = getFirstVisibleRow() > maxFirstRow;
-            boolean isOutOfColBounds =  getFirstVisibleColumn() > maxFirstCol;
+            boolean isOutOfColBounds = getFirstVisibleColumn() > maxFirstCol;
             if (isOutOfRowBounds || isOutOfColBounds) {
                 int firstRow;
                 if (isOutOfRowBounds) {
@@ -299,7 +302,7 @@ public class FixedGridLayoutManager extends RecyclerView.LayoutManager {
         }
 
 
-        mVisibleRowCount = (getVerticalSpace()/ mDecoratedChildHeight) + 1;
+        mVisibleRowCount = (getVerticalSpace() / mDecoratedChildHeight) + 1;
         if (getVerticalSpace() % mDecoratedChildHeight > 0) {
             mVisibleRowCount++;
         }
@@ -348,7 +351,7 @@ public class FixedGridLayoutManager extends RecyclerView.LayoutManager {
             }
 
             //Cache all views by their existing position, before updating counts
-            for (int i=0; i < getChildCount(); i++) {
+            for (int i = 0; i < getChildCount(); i++) {
                 int position = positionOfIndex(i);
                 final View child = getChildAt(i);
                 viewCache.put(position, child);
@@ -356,7 +359,7 @@ public class FixedGridLayoutManager extends RecyclerView.LayoutManager {
 
             //Temporarily detach all views.
             // Views we still need will be added back at the proper index.
-            for (int i=0; i < viewCache.size(); i++) {
+            for (int i = 0; i < viewCache.size(); i++) {
                 detachView(viewCache.valueAt(i));
             }
         }
@@ -476,7 +479,7 @@ public class FixedGridLayoutManager extends RecyclerView.LayoutManager {
          * that we did not re-attach. These are views that are not currently
          * necessary because they are no longer visible.
          */
-        for (int i=0; i < viewCache.size(); i++) {
+        for (int i = 0; i < viewCache.size(); i++) {
             final View removingView = viewCache.valueAt(i);
             recycler.recycleView(removingView);
         }
@@ -491,7 +494,7 @@ public class FixedGridLayoutManager extends RecyclerView.LayoutManager {
     @Override
     public void scrollToPosition(int position) {
         if (position >= getItemCount()) {
-            Log.e(TAG, "Cannot scroll to "+position+", item count is "+getItemCount());
+            Log.e(TAG, "Cannot scroll to " + position + ", item count is " + getItemCount());
             return;
         }
 
@@ -512,7 +515,7 @@ public class FixedGridLayoutManager extends RecyclerView.LayoutManager {
     @Override
     public void smoothScrollToPosition(RecyclerView recyclerView, RecyclerView.State state, final int position) {
         if (position >= getItemCount()) {
-            Log.e(TAG, "Cannot scroll to "+position+", item count is "+getItemCount());
+            Log.e(TAG, "Cannot scroll to " + position + ", item count is " + getItemCount());
             return;
         }
 
@@ -566,7 +569,7 @@ public class FixedGridLayoutManager extends RecyclerView.LayoutManager {
         //Take leftmost measurements from the top-left child
         final View topView = getChildAt(0);
         //Take rightmost measurements from the top-right child
-        final View bottomView = getChildAt(mVisibleColumnCount-1);
+        final View bottomView = getChildAt(mVisibleColumnCount - 1);
 
         //Optimize the case where the entire data set is too small to scroll
         int viewSpan = getDecoratedRight(bottomView) - getDecoratedLeft(topView);
@@ -647,7 +650,7 @@ public class FixedGridLayoutManager extends RecyclerView.LayoutManager {
         //Take top measurements from the top-left child
         final View topView = getChildAt(0);
         //Take bottom measurements from the bottom-right child.
-        final View bottomView = getChildAt(getChildCount()-1);
+        final View bottomView = getChildAt(getChildCount() - 1);
 
         //Optimize the case where the entire data set is too small to scroll
         int viewSpan = getDecoratedBottom(bottomView) - getDecoratedTop(topView);
@@ -726,7 +729,7 @@ public class FixedGridLayoutManager extends RecyclerView.LayoutManager {
      */
     @Override
     public View findViewByPosition(int position) {
-        for (int i=0; i < getChildCount(); i++) {
+        for (int i = 0; i < getChildCount(); i++) {
             if (positionOfIndex(i) == position) {
                 return getChildAt(i);
             }
@@ -735,7 +738,9 @@ public class FixedGridLayoutManager extends RecyclerView.LayoutManager {
         return null;
     }
 
-    /** Boilerplate to extend LayoutParams for tracking row/column of attached views */
+    /**
+     * Boilerplate to extend LayoutParams for tracking row/column of attached views
+     */
 
     /*
      * Even without extending LayoutParams, we must override this method
@@ -748,10 +753,12 @@ public class FixedGridLayoutManager extends RecyclerView.LayoutManager {
                 ViewGroup.LayoutParams.WRAP_CONTENT,
                 ViewGroup.LayoutParams.WRAP_CONTENT);
     }
+
     @Override
     public RecyclerView.LayoutParams generateLayoutParams(Context c, AttributeSet attrs) {
         return new LayoutParams(c, attrs);
     }
+
     @Override
     public RecyclerView.LayoutParams generateLayoutParams(ViewGroup.LayoutParams lp) {
         if (lp instanceof ViewGroup.MarginLayoutParams) {
@@ -760,6 +767,7 @@ public class FixedGridLayoutManager extends RecyclerView.LayoutManager {
             return new LayoutParams(lp);
         }
     }
+
     @Override
     public boolean checkLayoutParams(RecyclerView.LayoutParams lp) {
         return lp instanceof LayoutParams;
@@ -775,21 +783,27 @@ public class FixedGridLayoutManager extends RecyclerView.LayoutManager {
         public LayoutParams(Context c, AttributeSet attrs) {
             super(c, attrs);
         }
+
         public LayoutParams(int width, int height) {
             super(width, height);
         }
+
         public LayoutParams(ViewGroup.MarginLayoutParams source) {
             super(source);
         }
+
         public LayoutParams(ViewGroup.LayoutParams source) {
             super(source);
         }
+
         public LayoutParams(RecyclerView.LayoutParams source) {
             super(source);
         }
     }
 
-    /** Animation Layout Helpers */
+    /**
+     * Animation Layout Helpers
+     */
 
     /* Helper to obtain and place extra appearing views */
     private void layoutAppearingViews(RecyclerView.Recycler recycler, View referenceView, int referencePosition, int extraCount, int offset) {
@@ -854,12 +868,15 @@ public class FixedGridLayoutManager extends RecyclerView.LayoutManager {
                 layoutTop + mDecoratedChildHeight);
     }
 
-    /** Private Helpers and Metrics Accessors */
+    /**
+     * Private Helpers and Metrics Accessors
+     */
 
     /* Return the overall column index of this position in the global layout */
     private int getGlobalColumnOfPosition(int position) {
         return position % mTotalColumnCount;
     }
+
     /* Return the overall row index of this position in the global layout */
     private int getGlobalRowOfPosition(int position) {
         return position / mTotalColumnCount;
